make goreportcard happy
