+++ cli.go
@@ -290,136 +290,7 @@ func (p *App) printSuggestions(invalidCmdName string) {
 }
 
 func (p *App) printUsage() {
-	globalFlags := p.getGlobalFlags()
-	ctx := p.getParsingContext()
-
-	fs := ctx.getFlagSet()
-	out := getFlagSetOutput(fs)
-
-	if ctx.opts.customUsage != nil {
-		help := strings.TrimSpace(ctx.opts.customUsage())
-		fmt.Fprintf(out, "%s\n\n", help)
-		return
-	}
-
-	if !ctx.parsed && globalFlags != nil {
-		wrapArgs := &withGlobalFlagArgs{
-			GlobalFlags: globalFlags,
-		}
-		err := ctx.parseTags(reflect.ValueOf(wrapArgs).Elem())
-		if err != nil {
-			return
-		}
-	}
-
-	cmds := p.cmds
-	keepCmdOrder := p.opts.KeepCommandOrder
-
-	cmdName := ctx.name
-	flags := ctx.flags
-	nonflags := ctx.nonflags
-	showHidden := ctx.showHidden
-
-	flagCount := 0
-	hasShortFlag := false
-	for _, f := range flags {
-		if !f.hidden || showHidden {
-			flagCount++
-			hasShortFlag = hasShortFlag || f.short != ""
-		}
-	}
-	subCmds := cmds.listSubCommandsToPrint(cmdName, showHidden)
-	progName := getProgramName()
-	hasFlags, hasNonflags := flagCount > 0, len(nonflags) > 0
-	hasSubCmds := len(subCmds) > 0
-
-	usage := ""
-	cmd := ctx.cmd
-	if cmd != nil {
-		if cmd.AliasOf != "" {
-			usage += cmd.Description + "\n"
-			cmd = p.cmdMap[cmd.AliasOf]
-			cmdName = cmd.Name
-		}
-		if cmd.Description != "" {
-			usage += cmd.Description + "\n"
-		}
-		if usage != "" {
-			usage += "\n"
-		}
-	}
-	usage += "USAGE:\n  " + progName
-	if cmdName != "" {
-		usage += " " + cmdName
-	}
-	if hasFlags {
-		usage += " [flags]"
-	}
-	if hasNonflags {
-		for _, f := range nonflags {
-			name := f.name
-			if f.isSlice() {
-				name += "..."
-			} else if f.isMap() {
-				name += "{...}"
-			}
-			if f.required {
-				usage += fmt.Sprintf(" <%s>", name)
-			} else {
-				usage += fmt.Sprintf(" [%s]", name)
-			}
-		}
-	}
-	if !hasFlags && !hasNonflags && hasSubCmds {
-		usage += " <command> ..."
-	}
-	fmt.Fprint(out, usage, "\n\n")
-
-	if hasSubCmds {
-		printSubCommands(out, subCmds, showHidden, keepCmdOrder)
-		fmt.Fprint(out, "\n")
-	}
-
-	var globalFlagHelp [][2]string
-	var cmdFlagHelp [][2]string
-	if flagCount > 0 {
-		for _, f := range flags {
-			if f.hidden && !showHidden {
-				continue
-			}
-			name, usage := f.getUsage(hasShortFlag)
-			if f.isGlobal {
-				globalFlagHelp = append(globalFlagHelp, [2]string{name, usage})
-			} else {
-				cmdFlagHelp = append(cmdFlagHelp, [2]string{name, usage})
-			}
-		}
-	}
-	if len(cmdFlagHelp) > 0 {
-		fmt.Fprint(out, "FLAGS:\n")
-		printWithAlignment(out, cmdFlagHelp)
-		fmt.Fprint(out, "\n")
-	}
-	if len(nonflags) > 0 {
-		var nonflagLines [][2]string
-		for _, f := range nonflags {
-			name, usage := f.getUsage(false)
-			nonflagLines = append(nonflagLines, [2]string{name, usage})
-		}
-		fmt.Fprint(out, "ARGUMENTS:\n")
-		printWithAlignment(out, nonflagLines)
-		fmt.Fprint(out, "\n")
-	}
-	if len(globalFlagHelp) > 0 {
-		fmt.Fprint(out, "GLOBAL FLAGS:\n")
-		printWithAlignment(out, globalFlagHelp)
-		fmt.Fprint(out, "\n")
-	}
-
-	if ctx.opts.helpFooter != nil {
-		footer := strings.TrimSpace(ctx.opts.helpFooter())
-		fmt.Fprintf(out, "%s\n\n", footer)
-	}
+	newUsagePrinter(p).Do()
 }
 
 func printSubCommands(out io.Writer, cmds commands, showHidden bool, keepCmdOrder bool) {
